new base for vm
