+++ bot.py
@@ -14,7 +14,7 @@ intents.members = True
 
 bot = commands.Bot(command_prefix='.', intents=intents)  # bot command
 #coglist = ['dump','snail', 'help', 'dice', 'echo', 'pingstats', 'anime','dnd','uwu','modabuse','muri']
-coglist = ['rolls']
+coglist = ['anime','fx']
 #coglist = ['dump','snail', 'help', 'dice', 'echo', 'pingstats', 'anime','dnd','uwu','modabuse','muri']
 bot.remove_command('help')
 spook = [   "This only makes me stronger.",
