+++ cogs/old/rolls.py
